removed redundancy in scss files
