+++ src/components/Menu/Menu.scss
@@ -1,33 +1,22 @@
 .usa-button.menuBtn, .usa-button.menuBtn:hover {
+  background-color: #f0f0f0;
   width: 36px;
   height: 36px;
   border-radius: 60px;
-  border-width: 2px;
   border-style: outset;
-  margin: 0px;
-  padding: 0px;
-  background-color: #f0f0f0;
+  border-width: 2px;
   box-shadow: 0px 0px 2px #1488e7;
-  opacity: 1;
   z-index: 10;
   color: #1488e7;
-  border-color: #1488e7;
-  float: right !important;
 }
 
 .usa-nav__submenu.menuContent {
-  float: right;
-  position: absolute;
   bottom: 5px;
   right: 50px;
   width: 200px;
-  height: 120px;
-  z-index: auto;
-  text-align: left;
   box-shadow: 0px -2px 22px -6px rgba(60, 55, 61, 0.37),
     -10px 10px 22px -6px rgba(60, 55, 61, 0.37);
   background: #f1f1f1 0% 0% no-repeat padding-box;
-  padding: 15px;
   a {
     color: #0071bc !important;
     font-size: 0.7em !important;
@@ -48,7 +37,6 @@
   text-decoration: none;
   display: block;
   border-bottom: 1px solid lightgrey;
-  font: normal bold 15px Source Sans Pro, sans-serif;
 }
 
 .usa-nav__submenu-item:first-child {
