--- src/components/Menu/menuTopics.js
@@ -1,14 +1,0 @@
-export const collapsableFooterTopics = [
-  {
-    link: 'https://www.epa.gov/accessibility',
-    name: 'Accessibility',
-  },
-  {
-    link: 'https://www.epa.gov/privacy',
-    name: 'Privacy',
-  },
-  {
-    link: 'https://www.epa.gov/privacy/privacy-and-security-notice',
-    name: 'Privacy and Security Notice',
-  },
-];
