+++ src/utils/constants/menuTopics.js
@@ -134,3 +134,18 @@ export const aboutEPATopics = [
     name: "Labs and Research Centers",
   },
 ];
+
+export const collapsableFooterTopics = [
+  {
+    link: 'https://www.epa.gov/accessibility',
+    name: 'Accessibility',
+  },
+  {
+    link: 'https://www.epa.gov/privacy',
+    name: 'Privacy',
+  },
+  {
+    link: 'https://www.epa.gov/privacy/privacy-and-security-notice',
+    name: 'Privacy and Security Notice',
+  },
+];
